now properly fixed mistake
